refacto + some tests
