+++ gym_buster/__init__.py
@@ -5,8 +5,5 @@ logger = logging.getLogger(__name__)
 
 register(
     id='Buster-v0',
-    entry_point='gym_buster.envs:BusterEnv',
-    timestep_limit=1000,
-    reward_threshold=1.0,
-    nondeterministic=True,
+    entry_point='gym_buster.envs:BusterEnv'
 )
