+++ gym_buster/envs/game_classes/constants.py
@@ -1,0 +1,59 @@
+class Constants:
+    """
+    Class that will handle all constants needed
+    """
+    # MAP Coonstants
+    MAP_WIDTH = 16000
+    MAP_HEIGHT = 9000
+    MAP_MAX_DISTANCE = 99999
+    MAP_EMPTY_CELL = 0
+
+    # ENTITIES Constants
+    ENTITY_ID_LENGTH = 16
+    ENTITY_RANGE_VISION = 2200
+
+    GHOST_NUMBER = 9
+    BUSTER_NUMBER_PER_TEAM = 3
+    GAME_ROUND_NUMBER = 250
+
+    BASE_RANGE_GHOST_VALIDATED = 1600
+
+    GHOST_RUN_WAY = 220
+    BUSTER_MAX_MOVE = 800
+    BUSTER_BUST_MAX_RANGE = 1760
+    BUSTER_BUST_MIN_RANGE = 900
+
+    ACTION_MOVE_REGEX = r'MOVE (\d{0,5}) (\d{0,5})'
+    ACTION_RELEASE_REGEX = r'RELEASE'
+    ACTION_BUST_REGEX = r'BUST (\d+)'
+
+    # TYPE
+    TYPE_GHOST = -1
+    TYPE_BUSTER_TEAM_0 = 0
+    TYPE_BUSTER_TEAM_1 = 1
+
+    # STATE
+    STATE_GHOST = 0
+    STATE_BUSTER_CARRYING = 1
+    STATE_BUSTER_NOTHING = 0
+
+    # VALUE
+    VALUE_GHOST_BASIC = 0
+    VALUE_BUSTER_NOTHING = -1
+
+    # Rendering constants
+    FPS = 30
+
+    PYGAME_WINDOW_HEIGHT = 450  # pixels
+    PYGAME_WINDOW_WIDTH = 800  # pixels
+    PYGAME_WHITE = (255, 255, 255)
+    PYGAME_GHOST_COLOR = (255, 255, 255)
+    PYGAME_BUSTER_TEAM_0_COLOR = (0, 255, 0)
+    PYGAME_BUSTER_TEAM_1_COLOR = (255, 0, 0)
+    PYGAME_BLACK = (0, 0, 0)
+    PYGAME_GHOST_RADIUS = 10  # pixels
+    PYGAME_BUSTER_RADIUS = 10  # pixels
+    PYGAME_SPEED = 20  # pixels
+
+    PYGAME_RATIO_HEIGHT = PYGAME_WINDOW_HEIGHT / MAP_HEIGHT
+    PYGAME_RATIO_WIDTH = PYGAME_WINDOW_WIDTH / MAP_WIDTH
