+++ gym_buster/envs/game_classes/tests/_trial_temp-1/test.log
@@ -1,0 +1,1 @@
+2019-07-24 19:34:46+0200 [-] Log opened.
