+++ gym_buster/envs/game_classes/tests/_trial_temp-3/test.log
@@ -1,0 +1,1 @@
+2019-07-24 20:08:41+0200 [-] Log opened.
