+++ gym_buster/envs/game_classes/tests/_trial_temp-4/test.log
@@ -1,0 +1,1 @@
+2019-07-24 20:10:56+0200 [-] Log opened.
