+++ gym_buster/envs/game_classes/tests/_trial_temp/test.log
@@ -1,0 +1,1 @@
+2019-07-24 19:19:55+0200 [-] Log opened.
